move redis-based storage to its own package
