+++ pkg/storage/redis/config.go
@@ -1,4 +1,4 @@
-package storage
+package redis
 
 import (
 	"strings"
